fix: flight tracing
